pwrutilx: Avoid using invalid next_freq
commit 97739501f207efe33145b918817f305b822987f8 upstream.

If the next_freq field of struct sugov_policy is set to UINT_MAX,
it shouldn't be used for updating the CPU frequency (this is a
special "invalid" value), but after commit b7eaf1aab9f8 (cpufreq:
schedutil: Avoid reducing frequency of busy CPUs prematurely) it
may be passed as the new frequency to sugov_update_commit() in
pwrgov_update_single().

Fix that by adding an extra check for the special UINT_MAX value
of next_freq to pwrgov_update_single().